+++ lib/extensions/invitation/phoenix/templates/invitation_template.ex
@@ -21,7 +21,7 @@ defmodule PowInvitation.Phoenix.InvitationTemplate do
     {:password, :password_confirmation}
   ]) %>
 
-  <span><%%= link "Sign in", to: Routes.<%= Pow.Phoenix.Controller.route_helper(Pow.Phoenix.SessionController) %>_path(@conn, :new) %></span>
+  <span><%%= link "Sign in", to: <%= Pow.Phoenix.Template.__inline_route__(Pow.Phoenix.SessionController, :new) %>%></span>
   """
 
   template :show, :html,