+++ lib/pow/phoenix/templates/registration_template.ex
@@ -13,7 +13,7 @@ defmodule Pow.Phoenix.RegistrationTemplate do
   ],
   button_label: "Register") %>
 
-  <span><%%= link "Sign in", to: Routes.<%= Pow.Phoenix.Controller.route_helper(Pow.Phoenix.SessionController) %>_path(@conn, :new) %></span>
+  <span><%%= link "Sign in", to: <%= Pow.Phoenix.Template.__inline_route__(Pow.Phoenix.SessionController, :new) %>%></span>
   """
 
   template :edit, :html,